requirements.txt
langchain>=0.3.14
langchain-core>=0.3.30
langchain-ollama>=0.2.2
langchain_chroma
langchain-community>=0.3.14
ollama>=0.4.5
pyyaml>=6.0.2
#chromadb>=0.6.3
beautifulsoup4>=4.12.3
requests>=2.32.3